back-end/email_generator/requirements.txt
google-generativeai==0.3.0
python-dotenv==1.0.0
PyPDF2==3.0.1
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1